requirements.txt
asgiref==3.4.1
colorama==0.4.4
Django==3.2.5
gunicorn==20.1.0
numpy==1.21.0
opencv-python==4.5.2.54
Pillow==8.3.1
pytz==2021.1
qrcode==7.1
sqlparse==0.4.1
whitenoise==5.2.0
